+++ src/renderer/viz/expressions/RampGeneric.js
@@ -64,7 +64,7 @@ export default class RampGeneric extends Base {
         return mix(paletteValues[min], paletteValues[max], m);
     }
 
-    getLegend (options) {
+    getLegendData (options) {
         const config = Object.assign({}, DEFAULT_OPTIONS, options);
         const type = this.input.type;
         const legendData = this.input.getLegendData(config);